requirements.txt
coverage==4.5.3
djangorestframework==3.9.3
django==2.2.*
djangorestframework-simplejwt
python-decouple
django-jet==1.0.8
django-cors-headers
Pillow
django-rest-auth
django-rest-auth[with_social]
djangorestframework-jwt
jdatetime
psycopg2
markdown
django-redis
user-agents
celery
redis
django_extensions
gunicorn
